Fix ESLint errors causing CI failure
